Separate classes further into a command and task package
