+++ src/main/java/duke/Duke.java
@@ -1,5 +1,8 @@
 package duke;
 
+import command.Command;
+import task.TaskList;
+
 import java.io.FileNotFoundException;
 
 /**
