+++ src/main/java/duke/Parser.java
@@ -1,5 +1,18 @@
 package duke;
 
+import command.AddCommand;
+import command.ArchiveCommand;
+import command.Command;
+import command.DeleteCommand;
+import command.ExitCommand;
+import command.FindCommand;
+import command.ListCommand;
+import command.MarkCommand;
+import command.NullCommand;
+import command.ShowCommand;
+import command.UnmarkCommand;
+import task.*;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
