+++ src/main/java/duke/Storage.java
@@ -1,5 +1,11 @@
 package duke;
 
+import task.Deadline;
+import task.Event;
+import task.Task;
+import task.TaskList;
+import task.Todo;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
