+++ src/main/java/task/TaskList.java
@@ -1,4 +1,4 @@
-package duke;
+package task;
 
 import java.util.ArrayList;
 import java.util.List;
