+++ src/test/java/duke/TodoTest.java
@@ -1,22 +1,23 @@
 package duke;
 
 import org.junit.jupiter.api.Test;
+import task.Task;
+import task.Todo;
 
-import static duke.DukeConstants.KEY_SEPARATOR;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class TodoTest {
     @Test
     public void formatToSave_marked() {
         Task task = new Todo("test");
-        task.isDone = true;
+        task.markAsDone();
         String expectedStr = "T//1//test";
         assertEquals(expectedStr, task.formatToSave());
     }
     @Test
     public void formatToSave_unmarked() {
         Task task = new Todo("test");
-        task.isDone = false;
+        task.markAsUndone();
         String expectedStr = "T//0//test";
         assertEquals(expectedStr, task.formatToSave());
     }
