+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,3 @@
-hi im chompers what do u need!!!
-
 oke i added:
 [T][ ] submit hw
 now u have 1 task(s)!
