+++ text-ui-test/input.txt
@@ -1,1 +1,1 @@
-deadline submit hw /by friday
+todo submit hw
